Working prototype
First working prototype of Game Of Life. The algorithm is buggy (the right-hand part of the grid behaves strangely), but the basic functionality is there. The back-end code was imported and adapted from a homologous example developed to be run on a terminal. It is necessary to refactor and remove the numerous superfluous parts, which in the original were mainly used to print the grid on the terminal.

The structure of the program still needs to be described.