Simplified persistency
The persistency mechanism has been simplified: now that the `Backend` is using `QVector`s and `QScopedPointer`s, resizing the game board is just a matter of resizing a `QVector`.

All occurrences of "reinit" or "reinitialise" have been replaced by "resizing", because the `GameBoard` is not being destroyed and recreated any longer at every new game (this was a legacy from the previous management with `Cell**`).

Similarly, all occurrences of `clearHasJustBeenToggledFlag` or similar have been replaced with a simpler `unlockToggling` for consistency with the choice of identifiers.

`backend.hpp`
`backend.cpp`
- `resizeGameBoard` now delegates the actual resizing to `m_gameBoard`.

`gameboard.hpp`
`gameboard.cpp`
- Method `resizeGameBoard` added.