Simplification of previous commit
It turned out that it was not necessary to relay information through the `GameManager`. As a matter of fact, the children objects of a QML object can recognise an `id` declared in a parent object, even if the children (or grandchildren) are defined in another QML document.

`GameBoard.qml`
- There is no need to inform the `GameManager` about the status of the game timer.

`GameCell.qml`
- As for the `GameBoard.qml`, there is no need to inform the `GameManager` about the status of the game timer. The `timer` identifier is recognised by `GameCell.qml`, even though `timer` is defined in `GameBoard.qml`, which is the "grandfather" of `GameCell.qml` (`GameBoard.qml` declares a `CustomScrollView.qml` which in turn declares a `GameCell.qml`).

`gamemanager.hpp`
`gamemanager.cpp`
- These two files have been reverted to their previous status, i.e., no mentioning of the game timer.